Further clarify/constrain the cue schema for Project
This PR constrains how "open" the `administrators` and
`repositories` list types are defined in the schema.

With their previous definitions, the following project
was valid according to the schema:

```yaml
project:
  name: example
  administrators: []
  repositories: []
  vulnerability-reporting:
    reports-accepted: true
    bug-bounty-available: true
```

Given that the `template-minimum.yml` already illustrates
this constraint and it makes logical sense, it is included
in the cue schema and the Markdown docs for Project are
updated to note the required list size of 1 for each list.

Signed-off-by: Travis Truman <trumant@gmail.com>