docs: add adoption guidance docs and example files (#130)
This change updates `README.md` to provide greater
clarity to projects adopting insights and polishes
some existing guidance that was inaccurate/out-of-date.

The new example files provided extend the guidance
in the README and are referred to within that guidance.

This closes https://github.com/ossf/security-insights-spec/issues/126

This updates https://github.com/ossf/security-baseline/issues/257

Signed-off-by: Travis Truman <trumant@gmail.com>